requirements.txt
# Core AI/ML libraries
torch>=1.9.0
transformers>=4.21.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
openai>=1.3.0

# PDF processing
PyMuPDF>=1.23.0

# Image processing
Pillow>=9.0.0

# Web interface
gradio>=4.0.0

# Utilities
python-dotenv>=0.19.0
numpy>=1.21.0
scikit-learn>=1.0.0

# Optional: For GPU support (uncomment if you have CUDA)
# faiss-gpu>=1.7.0
# torch-audio>=0.9.0